ref(processor): Create event processor sub-module with related code
